% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectrum.R
\name{dp_read_anc_testing}
\alias{dp_read_anc_testing}
\alias{dp_read_pmtct}
\alias{dp_read_pmtct_retained}
\alias{dp_read_abortion}
\alias{dp_read_mothers_reallocated}
\alias{dp_read_breastfeeding}
\alias{dp_read_childart}
\alias{dp_read_childltfu}
\alias{dp_read_art_dist}
\alias{dp_read_child_mort_mult}
\alias{dp_read_nosocom_infections}
\alias{dp_read_mtct_rates}
\alias{dp_read_paed_cd4_dist}
\alias{dp_read_paed_cd4_prog}
\alias{dp_read_paed_cd4_mort}
\alias{dp_read_paed_art_mort}
\alias{dp_read_paed_art_eligibility}
\title{Read Spectrum programme data inputs}
\usage{
dp_read_anc_testing(dp)

dp_read_pmtct(dp)

dp_read_pmtct_retained(dp)

dp_read_abortion(dp)

dp_read_mothers_reallocated(dp)

dp_read_breastfeeding(dp)

dp_read_childart(dp)

dp_read_childltfu(dp)

dp_read_art_dist(dp)

dp_read_child_mort_mult(dp)

dp_read_nosocom_infections(dp)

dp_read_mtct_rates(dp)

dp_read_paed_cd4_dist(dp)

dp_read_paed_cd4_prog(dp)

dp_read_paed_cd4_mort(dp)

dp_read_paed_art_mort(dp)

dp_read_paed_art_eligibility(dp)
}
\arguments{
\item{dp}{An \code{"spectrum_dp"} object created by \code{read_dp()} or a path to a Spectrum
PJNZ file.}
}
\value{
Most functions return a matrix in which rows are each indicator and columns are years

\code{dp_read_abortion()} returns a list in which the vector \code{pregtermabortion} is the
numerical values and \code{pregtermabortion_isperc} is a logical indicating whether each
value corresponds to a number (count) input or percentage input.
}
\description{
Read Spectrum programme data inputs
}
\details{
These functions accepts either a path to a PJNZ file or an already parsed \code{.DP}
file read using \code{\link[eppasm:read_dp]{eppasm::read_dp()}} with class \code{"spectrum_dp"}. Unzipping and reading the
\code{.DP} file is the slow part of the function, so if reading lots of data from
the same file, it will be most efficient to read it once and pass that to the
functions.
}
\examples{
\dontrun{
pjnz <- system.file(
  "pjnz/bwa_aim-adult-art-no-special-elig_v6.13_2022-04-18.PJNZ",
  package = "leapfrog")
dp <- read_dp(pjnz)
dp_anc_testing <- dp_read_anc_testing(dp)
dp_pmtct <- dp_read_pmtct(dp)
dp_read_pmtct <- dp_read_pmtct_retained(dp)
dp_abortion <- dp_read_abortion(dp)
dp_notbreastfeeding <- dp_read_breastfeeding(dp)
dp_childart <- dp_read_childart(dp)
dp_childltfu <- dp_read_childltfu(dp)

## Can either pass PJNZ path or parsed "spectrum_dp" object

anc_testing1 <- dp_read_anc_testing(pjnz)
anc_testing2 <- dp_read_anc_testing(dp)
all.equal(anc_testing1, anc_testing2)
}
}
\keyword{internal}
