% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_leapfrog_ss}
\alias{get_leapfrog_ss}
\title{Get the state space dimensions for a particular model configuration}
\usage{
get_leapfrog_ss(configuration)
}
\arguments{
\item{configuration}{The configuration to get state space for,
see \link{list_model_configurations}.}
}
\value{
State space as a named list
}
\description{
Get the state space dimensions for a particular model configuration
}
