% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{get_state_space}
\alias{get_state_space}
\title{Run leapfrog model fit}
\usage{
get_state_space(configuration = "HivFullAgeStratification")
}
\arguments{
\item{configuration}{The model configuration to run, see
\code{\link[=list_model_configurations]{list_model_configurations()}} for available configurations}
}
\value{
The state space for this model configuration.
}
\description{
Run leapfrog model fit
}
\examples{
s <- get_state_space("HivCoarseAgeStratification")
}
