% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_model_configurations}
\alias{list_model_configurations}
\title{List the avaialble model configurations}
\usage{
list_model_configurations()
}
\value{
List of available model configurations
}
\description{
List the avaialble model configurations
}
