% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectrum.R
\name{prepare_hc_leapfrog_projp}
\alias{prepare_hc_leapfrog_projp}
\title{Prepare child HIV projection parameters from Spectrum PJNZ}
\usage{
prepare_hc_leapfrog_projp(pjnz, params, use_coarse_age_groups = FALSE)
}
\arguments{
\item{pjnz}{path to PJNZ file}

\item{params}{The adult HIV projection parameters from \code{\link[=prepare_leapfrog_projp]{prepare_leapfrog_projp()}}}
}
\value{
List of child HIV projection parameters
}
\description{
Prepare child HIV projection parameters from Spectrum PJNZ
}
\examples{
pjnz <- system.file(
  "pjnz/bwa_aim-adult-art-no-special-elig_v6.13_2022-04-18.PJNZ",
  package = "leapfrog", mustWork = TRUE)
demp <- prepare_leapfrog_demp(pjnz)
proj <- prepare_leapfrog_projp(pjnz)
params <- c(proj, demp)
projp <- prepare_hc_leapfrog_projp(pjnz, params)
}
