% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_inputs.R
\name{prepare_leapfrog_projp}
\alias{prepare_leapfrog_projp}
\title{Prepare adult HIV projection parameters from Spectrum PJNZ}
\usage{
prepare_leapfrog_projp(
  pjnz,
  use_coarse_age_groups = FALSE,
  hiv_steps_per_year = 10L,
  hTS = 3
)
}
\arguments{
\item{pjnz}{path to PJNZ file}

\item{hiv_steps_per_year}{number of Euler integration steps per year; default 10}

\item{hTS}{number of HIV treatment stages; default 3 (0-5 months,
6-11 months, 12+ months)}
}
\value{
list of HIV projection parameters
}
\description{
Prepare adult HIV projection parameters from Spectrum PJNZ
}
\examples{
pjnz <- system.file(
  "pjnz/bwa_aim-adult-art-no-special-elig_v6.13_2022-04-18.PJNZ",
  package = "leapfrog")
projp <- prepare_leapfrog_projp(pjnz)

}
