% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pjnz.R
\name{process_pjnz}
\alias{process_pjnz}
\title{Prepare inputs from Spectrum PJNZ}
\usage{
process_pjnz(pjnz, use_coarse_age_groups = FALSE, bypass_adult = FALSE)
}
\arguments{
\item{pjnz}{path to PJNZ file}

\item{use_coarse_age_groups}{use the coarse age stratification}

\item{bypass_adult}{produce parameters that will bypass the adult
model when running the child model variant}
}
\value{
list of input parameters
}
\description{
Prepare inputs from Spectrum PJNZ
}
\examples{
pjnz <- system.file(
  "pjnz/bwa_aim-adult-art-no-special-elig_v6.13_2022-04-18.PJNZ",
  package = "leapfrog")
parameters <- process_pjnz(pjnz)

}
