% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model_from_state}
\alias{run_model_from_state}
\title{Run leapfrog model fit from initial state}
\usage{
run_model_from_state(
  parameters,
  configuration,
  initial_state,
  simulation_start_year,
  output_years = seq(1970, 2030)
)
}
\arguments{
\item{parameters}{Projection parameters}

\item{configuration}{The model configuration to run, see
\code{\link[=list_model_configurations]{list_model_configurations()}} for available configurations}

\item{initial_state}{The model will run from this initial state}

\item{simulation_start_year}{Start the model simulation from a particular year}

\item{output_years}{Which years of the model to return from the simulation,
defaults to all years from 1970 to 2030. Also used to control what years
the simulation is run for. If output only 2030, simulation will be run
from \code{projection_start_year} passed in the \code{parameters} list.}
}
\value{
List of model outputs, where the last dimension of each element is
time, e.g. \code{p_totpop} state variable has dimensions 81 x 2. If
\code{output_years} specified has length 61 then the \code{p_totpop} output
will have dimensions 81 x 2 x 61.
}
\description{
Run leapfrog model fit from initial state
}
\examples{
pjnz <- system.file(
  "pjnz/bwa_aim-adult-art-no-special-elig_v6.13_2022-04-18.PJNZ",
  package = "leapfrog", mustWork = TRUE)
parameters <- prepare_leapfrog_parameters(pjnz, use_coarse_age_groups = TRUE)
out_first_half_years <- run_model(parameters, "HivCoarseAgeStratification", 1970:2000)
out_second_half_years <- run_model_from_state(
  parameters,
  "HivCoarseAgeStratification",
  get_time_slice(out_first_half_years, 31),
  2000,
  2001:2030)
}
